Свойство Difficulty хранит число меньшего обьёма чем int, поэтому для экономии памяти достаточно иметь тип byte

+++ Basket and stones/Assets/Scripts/MainMenu.cs
@@ -11,7 +11,7 @@ public class MainMenu : MonoBehaviour, IPhoneButtons
     public GameObject SettingsWindow, MainMenuWindow, DifficultyWindow;
     public Text DifficultLevelLabel;
     public Slider DifficultSlider;
-    public static int Difficulty;
+    public static byte Difficulty;
   
 
     public void OnMouseUpAsButton()
